+++ docs/src/components/container/container.types.ts
@@ -1,0 +1,3 @@
+import { ParentProps } from "solid-js";
+
+export type ContainerProps = ParentProps & {};